models: now models exchange global id
Now, each model exchange Id in output and external transition functions.
The template Value is now an integer.

+++ defs.hpp
@@ -42,7 +42,7 @@ struct Infinity
 };
 
 typedef vle::Time <double, Infinity<double>> Time;
-typedef double Data;
+typedef int Data;
 
 typedef vle::dsde::Engine <Time, Data> DSDE;
 typedef vle::dsde::Factory <Time, Data> Factory;
